refactor: extract vite plugins into its own package
