+++ frontend/package.json
@@ -1,7 +1,6 @@
 {
   "name": "@jellyfin-vue/frontend",
   "version": "0.3.1",
-  "private": true,
   "type": "module",
   "browserslist": [
     "defaults and supports es6-module"
@@ -53,6 +52,7 @@
     "@iconify/json": "2.2.246",
     "@intlify/eslint-plugin-vue-i18n": "3.0.0",
     "@intlify/unplugin-vue-i18n": "4.0.0",
+    "@jellyfin-vue/vite-plugins": "*",
     "@rollup/plugin-virtual": "3.0.2",
     "@stylistic/eslint-plugin": "2.8.0",
     "@types/dompurify": "3.0.5",
@@ -62,7 +62,6 @@
     "@unocss/eslint-config": "0.62.3",
     "@vitejs/plugin-vue": "5.1.3",
     "browserslist": "4.23.3",
-    "eslint": "9.10.0",
     "eslint-config-flat-gitignore": "0.3.0",
     "eslint-plugin-css": "0.11.0",
     "eslint-plugin-depend": "0.11.0",
@@ -78,8 +77,6 @@
     "eslint-plugin-vue-scoped-css": "2.8.1",
     "globals": "15.9.0",
     "lightningcss": "1.26.0",
-    "rollup-plugin-visualizer": "5.12.0",
-    "typescript": "5.6.2",
     "typescript-eslint": "8.5.0",
     "unocss": "0.62.3",
     "unplugin-icons": "0.19.3",
