+++ packages/vite-plugins/package.json
@@ -1,0 +1,10 @@
+{
+  "name": "@jellyfin-vue/vite-plugins",
+  "version": "0.0.1",
+  "exports": "./index.ts",
+  "type": "module",
+  "devDependencies": {
+    "rollup-plugin-visualizer": "5.12.0",
+    "vite": "5.4.3"
+  }
+}
